requirements.txt
selenium~=4.27.0
behave~=1.2.6
python-dotenv~=1.0.1
requests~=2.32.3
webdriver-manager~=4.0.2
allure-behave
behave-html-formatter
wheel>=0.36.2
pyperclip~=1.9.0
ua-generator
gspread
pandas~=2.2.3
openpyxl
certifi~=2025.1.31
mysql-connector-python~=9.3.0
tldextract~=5.3.0